RIPE NCC has merged bce9b25a6
* Clean a sonarqube nit [f2c12853c]
* Clean commandaudit in integration tests [0eec93b63]
* Fix getCAStatEvents [360db641a]
* JpaCertificateAuthorityRepository: fix class cast error [a5a732830]
* Add test cases for CaStat methods [9faca1eec]
* SecurityConfig: fix admin UI access when authorization is disabled [52ed3b2ad]
* SecurityConfig: fix provisioning config [cf384d82b]
* SecurityConfig: disable session request cache parameter [c126055eb]
* SecurityConfig: use java method naming conventions [19d678a61]
* OIDC needs to be added in the same chain that it applies to [5401517af]
* Rewrite the SecurityConfig [68edba4c6]
* SonarQube nits [98939469b]
* Fix the disabled test in UpStreamCARequestEntityTest [91937c1d8]
* Build a container with JRE 17 [1c0e6ff2e]
* Truncate instant to ms in ProvisioningCmsSigningTimeStoreTest [543a05e6d]
* Use Instant instead of Timestamp in JpaRoaConfigurationRepository [e69771564]
* Update query in JpaRoaConfigurationRepositoryTest for new types [aee572b62]
* Update shouldFindAllPerCaAndCountPrefixes for new types [9f26fcfc2]
* Two unnecessary imports [cd393d814]
* Adjust ExpireOutgoingResourceCertificateResult to use long [0c401f547]
* Use Instant in PublishedObject(Data|Entry) [1eee1854d]
* Hibernate returns Instant instead of Timestamp [cfe8b4e73]
* Result is a long with Hibernate 6 [2df77ebca]
* Store artefacts for failed builds [b8ca3ff78]
* Use lenient stubbing in UpstreamCaControllerTest [2eed0dc98]
* Use thymeleaf-spring6 and refactor admin tests to Jupiter APIs [3e27c1817]
* Fix incorrect JPQL reference to certificate ids [3edc720ab]
* Use Wiremock for integration test case [1b720adb7]
* CertificationDomainTestCase did not persist KeyPair [a9d90164d]
* Remove usage of RandomUtil [90d284304]
* Use new DefaulTransactionStatus constructor [da9d23a7f]
* Validate.notNull without explanation is deprecated [ec5593ef8]
* Use new Spring Security syntax [eb262c1a3]
* AutoConfigureMetrics is now AutoConfigureObservability [ec51a5ddb]
* Enum is now mapped as tinyint [d62fd4e14]
* Hibernate version specific dialects are deprecated [0c5b01391]
* javax.persistence -> jakarta.persistence [7b805d021]
* Switch back to DBProvider name [8a555792d]
* New jar artifact name [b84fa63c2]
* Test DBProvider 1.6-SNAPSHOT [e2117413f]
* additional .toList changes after rebase [b5371e0ad]
* Update default `JAVA_HOME` to pick java 17 [d90e2186f]
* Fix `toList` derived types [a54d1a8ce]
* Low-handing Java 17 changes [433da1eb1]
* Toolchain is needed once [2f47cb82f]
* Switch to Java 17 [de744b3b9]
* Use postgres 15 in unit tests run on Github actions [de72145a1]
* chore(deps): update plugin com.google.cloud.tools.jib to v3.4.2 [69c6cb91c]
* chore(deps): update dependency org.springdoc:springdoc-openapi-ui to v1.8.0 [59403bcb7]
* chore(deps): update dependency gradle to v8.7 [515ba2e4a]
* Remove commons-text and do not use tomcat dependency [4171caf49]
* chore(deps): update dependency commons-io:commons-io to v2.16.1 [f0512bbc8]
* chore(deps): update dependency org.postgresql:postgresql to v42.7.3 [255ed5d55]
* Remove workaround [b1c3f03b1]
* chore(deps): update dependency net.jqwik:jqwik to v1.8.4 [4eb03d2dc]

+++ src/main/java/net/ripe/rpki/domain/ResourceCertificateRepository.java
@@ -74,10 +74,10 @@ public interface ResourceCertificateRepository extends Repository<ResourceCertif
 
     @Value
     class ExpireOutgoingResourceCertificatesResult {
-        int expiredCertificateCount;
-        int deletedRoaCount;
-        int deletedAspaCount;
-        int withdrawnObjectCount;
+        long expiredCertificateCount;
+        long deletedRoaCount;
+        long deletedAspaCount;
+        long withdrawnObjectCount;
     }
 
     /**
